+++ app/(client)/(anon)/spots/[spotId]/tips/components/Reaction.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import React, { useState, useRef, useEffect } from "react";
+import { TipReactionDto } from "@/application/usecases/spot/dto/TipReactionDto";
 import Image from "next/image";
-import { PiSirenFill } from "react-icons/pi";
+import { useEffect, useRef, useState } from "react";
 import Emotion from "./Emotion";
 import styles from "./Reaction.module.scss";
-import { TipReactionDto } from "@/application/usecases/spot/dto/TipReactionDto";
+import Report from "./Report";
 
 const Reaction = ({
     tipReaction,
@@ -18,7 +18,6 @@ const Reaction = ({
 }) => {
     const [isModalOpen, setIsModalOpen] = useState(false);
     const [showMessage, setShowMessage] = useState(false);
-    const [errorMessage, setErrorMessage] = useState<string | null>(null);
     const modalRef = useRef<HTMLDivElement | null>(null);
 
     // 접속중인 유저 아이디
@@ -80,34 +79,6 @@ const Reaction = ({
         }
     };
 
-    const handleReportClick = async () => {
-        const confirm = window.confirm("신고하시겠습니까?");
-        if (!confirm) return;
-
-        const response = await fetch(
-            `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/tips/${tipId}/report`,
-            {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({
-                    userId,
-                }),
-            }
-        );
-        if (response.ok) {
-            setErrorMessage("신고가 완료되었습니다.");
-            setTimeout(() => {
-                setErrorMessage(null);
-            }, 1000);
-        } else {
-            setErrorMessage("이미 신고한 팁입니다.");
-            setTimeout(() => {
-                setErrorMessage(null);
-            }, 1000);
-        }
-    };
     useEffect(() => {
         if (isModalOpen) {
             document.addEventListener("mousedown", handleClickOutside);
@@ -134,9 +105,6 @@ const Reaction = ({
 
     return (
         <div className={styles.reactionContainer}>
-            {errorMessage && (
-                <div className={styles.errorMessage}>{errorMessage}</div>
-            )}
             <button
                 className={styles.reactionButton}
                 onClick={handleButtonClick}
@@ -190,9 +158,8 @@ const Reaction = ({
                     <Emotion count={typeCounts[4]} type={4} />
                 )}
             </div>
-            <button className={styles.sirenButton} onClick={handleReportClick}>
-                <PiSirenFill color="var(--red-1-color)" size={16} />
-            </button>
+
+            <Report tipId={tipId} userId={userId} />
         </div>
     );
 };
