+++ app/(client)/(anon)/spots/[spotId]/tips/components/Report.module.scss
@@ -1,0 +1,23 @@
+.reportButton {
+  background: none;
+  border: 0;
+  align-self: center;
+  width: 1rem;
+  height: 1rem;
+  margin-left: 1rem;
+  cursor: pointer;
+}
+
+.errorMessage {
+  position: fixed;
+  top: 95%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: rgba(235, 40, 40, 0.8);
+  color: black;
+  padding: 0.625rem 1.25rem;
+  border-radius: 1.25rem;
+  z-index: 1000;
+  width: 70%;
+  text-align: center;
+}
